+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-### 1.7.22 (build 21159, api 8, 2023-07-10)
+### 1.7.23 (build 21164, api 8, 2023-07-11)
+
+### 1.7.22 (build 21162, api 8, 2023-07-11)
 
 - Fixed a very rare race condition when launching threads or sending synchronous
   cross-thread messages. This was manifesting as one out of several thousand
@@ -11,6 +13,8 @@
   see any graphics wonkiness with this update.
 - Added SoK's explodinary icon to the game's custom text drawing because SoK is
   awesome.
+- (build 21161) Fixed an issue on Android that could lead to crashes if device
+  events occurred very early at launch (button presses, joystick movement, etc.)
 
 ### 1.7.21 (build 21152, api 8, 2023-06-27)